requirements.txt
aiogram>=3.3.0
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.4.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
redis>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
loguru>=0.7.2
fastapi>=0.109.0
uvicorn>=0.27.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
python-multipart>=0.0.9
